Created _cliutils module with utility stuff for CLI scripts.
Details:
- Created _cliutils.py. It will be picked up automatically by build.
- Moved the SmartFormatter class from wbemcli.py and mof_compiler.py
  into _cliutils.py.
- Changed `import pywbem` in wbemcli.py to `from . import WBEMConnection`
  because wbemcli is a pywbem module.
